SNOW-683060 fix mixup between unit/udtf.py and unit/udf.py
