+++ tests/unit/test_udf.py
@@ -2,7 +2,6 @@
 # Copyright (c) 2012-2022 Snowflake Computing Inc. All rights reserved.
 #
 
-from typing import Iterable, Tuple
 from unittest import mock
 
 import pytest
@@ -10,8 +9,9 @@ import pytest
 from snowflake.connector import ProgrammingError
 from snowflake.snowpark import Session
 from snowflake.snowpark.exceptions import SnowparkSQLException
-from snowflake.snowpark.functions import udtf
-from snowflake.snowpark.udtf import UDTFRegistration
+from snowflake.snowpark.functions import udf
+from snowflake.snowpark.types import IntegerType
+from snowflake.snowpark.udf import UDFRegistration
 
 
 @mock.patch("snowflake.snowpark.udf.cleanup_failed_permanent_registration")
@@ -21,24 +21,14 @@ def test_do_register_sp_negative(cleanup_registration_patch):
         return_value="database.schema"
     )
     fake_session._run_query = mock.Mock(side_effect=ProgrammingError())
-    fake_session.udtf = UDTFRegistration(fake_session)
+    fake_session.udf = UDFRegistration(fake_session)
     with pytest.raises(SnowparkSQLException) as ex_info:
-
-        @udtf(output_schema=["num"], session=fake_session)
-        class UDTFProgrammingErrorTester:
-            def process(self, n: int) -> Iterable[Tuple[int]]:
-                yield (n,)
-
+        udf(lambda: 1, session=fake_session, return_type=IntegerType(), packages=[])
     assert ex_info.value.error_code == "1304"
     cleanup_registration_patch.assert_called()
 
     fake_session._run_query = mock.Mock(side_effect=BaseException())
-    fake_session.udtf = UDTFRegistration(fake_session)
+    fake_session.udf = UDFRegistration(fake_session)
     with pytest.raises(BaseException):
-
-        @udtf(output_schema=["num"], session=fake_session)
-        class UDTFBaseExceptionTester:
-            def process(self, n: int) -> Iterable[Tuple[int]]:
-                yield (n,)
-
+        udf(lambda: 1, session=fake_session, return_type=IntegerType(), packages=[])
     cleanup_registration_patch.assert_called()
